% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heteroMatrix.R
\name{heteroMatrix}
\alias{heteroMatrix}
\title{Heterogenous Variance-Covariance Matrix}
\usage{
heteroMatrix(model)
}
\arguments{
\item{model}{An nlme model object}
}
\description{
This function computes a heterogenous variance-covariance matrix from an nlme model object.
}
\examples{
heteroMatrix(model)

}

