% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icc.R
\name{icc}
\alias{icc}
\title{Intraclass Correlation}
\usage{
icc(model)
}
\arguments{
\item{model}{A fitted nlme object}
}
\description{
This function computes an intraclass correlation value from
an empty random intercept fitted nlme model object. This function
is specific to 2-level models.
}
\examples{
icc(model1)

}

