% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudoR.R
\name{pseudoR}
\alias{pseudoR}
\title{Pseudo R-Squared}
\usage{
pseudoR(model1, model2, varParam)
}
\arguments{
\item{model1}{A fitted nlme object with fewer fixed effect
parameters than model B, but with an equal number of variance components.}

\item{model2}{A fitted nlme object with more fixed effect
parameters than model B, but with an equal number of variance components.}

\item{varParam}{The variance component that you are interested in
computing an pseudo R-squared for. That is, the variance component
that you expect to change due to adding a fixed effect.}
}
\description{
This function computes a pseudo R-squared value for the
amount of variance explained in a variance component
of nested nlme model objects. This function is specific to
2-level models. Note that model1 is nested in model2.
}
\examples{
randomCI(model1, "time")

}

