% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomCI.R
\name{randomCI}
\alias{randomCI}
\title{Random Effect Ninety-Five Percent Confidence Intervals}
\usage{
randomCI(model, Param)
}
\arguments{
\item{model}{A fitted nlme object}

\item{Param}{The random effect of interest. Use quotes.}
}
\description{
This function computes a random effect 95% confidence interval from
a fitted nlme object. That is, it computes an estimate of individual
differences (i.e., random effect variation) for a fixed effect. This
function is specific to 2-level models.
}
\examples{
randomCI(model1, "time")

}

