% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCsv.R
\name{readCsv}
\alias{readCsv}
\title{Read CSV File}
\usage{
readCsv(file, naString = NULL)
}
\arguments{
\item{file}{A csv file in your working directory. Use quotes.}

\item{naString}{An optional argument specifying one extra character/string/number to assign NA to.
Use quotes if it is a character or string. Only one extra character/string/number currently supported.}
}
\description{
This function reads a CSV file in a standard way.
Header is set to TRUE. stringsAsFactors is set to FALSE.
na.strings defaults to c("", " ", "NA", ".").
}
\examples{
readCsv(myFile.csv)

}

