% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreDass.R
\name{scoreDass}
\alias{scoreDass}
\title{Score DASS Scale}
\usage{
scoreDass(currWk, weekNum, courseName, masterFile)
}
\arguments{
\item{currWk}{A .csv file for the current week to score that you downloaded from Qualtrics.
Use quotes.}

\item{weekNum}{Of class character denoting the current week that you are going to score.
Use quotes.}

\item{courseName}{The course name formatted as follows:
locationBeginningAndEndingMonthofCourseYear, where the location is a 3 letter
abbreviation and the beginning and ending months are shortened to 3 letters if
longer than 5 letters and do not include follow-up months. Use quotes. See example.}

\item{masterFile}{A .csv file that includes computed DASS scores from all courses
of a specific type. If this is the first course of a specific class type that you
are scoring, I will create the file for you as long as you input a file name in
quotes with a .csv extension. Use quotes.}
}
\description{
I can be used to score a week's worth of DASS data that has been downloaded
from Qualtrics and either append it to a master.csv file or create a master.csv
file. If your working directory has a 'master' DASS file in it (e.g., masterCCTDass.csv)
where you are combining data from courses of the same class type but from different
locations, I will append this week to it and nest the data accordingly. If your working
directory does not have a masterCOURSENAMEDass.csv file in it, I will create that file
(to the name of your choosing) to start you off collecting data. I assume that there are
two character rows to remove below the header (i.e., rows 1 & 2). I assume that the
participant ID column is labeled 'ID' and that the DASS columns begin with 'DASS_.' I assume
that there are 16 columns to remove from the front of the Qualtrics CSV file.
Dependencies: dplyr and tidyr.
}
\examples{
scoreDass(currWk = "SacCCTWeek5A.csv", weekNum = "5A",
courseName = "sacAprilJune2016", masterFile = "masterCCTDass.csv")

}
