% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeSeriesChart.R
\name{timeSeriesChart}
\alias{timeSeriesChart}
\title{Time Series Chart}
\usage{
timeSeriesChart(data, timepoint, response, tx)
}
\arguments{
\item{data}{Is the dataframe that you are interested in visualizing.}

\item{timepoint}{Is the name of the column containing ordered time points.}

\item{response}{Is the name of the column containing the dependent (y axis) variable values.}

\item{tx}{Is optional and may be specified to create an interaction graph. It is the name
of the column containing treatment group membership.}
}
\description{
Creates a time series chart with data structured in long format. Dependenceis:
ggplot2 and plotly.
}
\examples{
timeSeriesChart(meanMcs, time, value, MBSR1CBT2)

}

